+++ README.md
@@ -1,4 +1,4 @@
-# forcats
+# catspy
 
 [![Build Status](https://travis-ci.org/MangoTheCat/forcats.svg?branch=master)](https://travis-ci.org/MangoTheCat/forcats)
