+++ catspy/__init__.py
@@ -1,0 +1,4 @@
+""" Tools for working with categorical data
+"""
+
+from catspy.functions import cat_lump, cat_count, cat_anon  # noqa
