--- forcats/__init__.py
@@ -1,4 +1,0 @@
-""" Tools for working with categorical data
-"""
-
-from forcats.functions import cat_lump, cat_count, cat_anon  # noqa
