+++ setup.py
@@ -4,11 +4,11 @@ with open("README.md", "r") as f:
     long_description = f.read()
 
 setup(
-    name="forcats",
+    name="catspy",
     version="0.1.0",
-    packages=["forcats"],
+    packages=["catspy"],
     description="Tools for working with categorical data",
-    url="http://github.com/MangoTheCat/forcats",
+    url="http://github.com/MangoTheCat/catspy",
     author="Mango Solutions",
     zip_safe=False,
     install_requires=["pandas", "numpy"],
